+++ src/openforms/api/views/__init__.py
@@ -1,0 +1,9 @@
+from .mixins import ListMixin
+from .views import ERR_CONTENT_TYPE, PingView, exception_handler
+
+__all__ = [
+    "ListMixin",
+    "ERR_CONTENT_TYPE",
+    "exception_handler",
+    "PingView",
+]
