+++ src/openforms/registrations/api/views.py
@@ -4,7 +4,7 @@ from drf_spectacular.utils import extend_schema, extend_schema_view
 from rest_framework import authentication, permissions
 from rest_framework.views import APIView
 
-from openforms.utils.api.views import ListMixin
+from openforms.api.views import ListMixin
 
 from ..constants import RegistrationAttribute
 from ..registry import register
